+++ package.json
@@ -6,14 +6,7 @@
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "web": "expo start --web",
-    "test": "jest"
-  },
-  "jest": {
-    "preset": "react-native",
-    "setupFilesAfterEnv": [
-      "@testing-library/jest-native/extend-expect"
-    ]
+    "web": "expo start --web"
   },
   "dependencies": {
     "@react-navigation/native": "^6.1.8",
@@ -29,11 +22,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
-    "@testing-library/jest-native": "^5.4.3",
-    "@testing-library/react-native": "^12.3.0",
-    "jest": "^29.7.0",
-    "react-native-dotenv": "^3.4.9",
-    "react-test-renderer": "^18.2.0"
+    "react-native-dotenv": "^3.4.9"
   },
   "private": true
 }
